added update method to TeacherService, added custom exceptions

+++ src/main/java/com/capgemini/gradebook/service/impl/TeacherServiceImpl.java
@@ -1,6 +1,7 @@
 package com.capgemini.gradebook.service.impl;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -13,6 +14,8 @@ import com.capgemini.gradebook.domain.mapper.TeacherMapper;
 import com.capgemini.gradebook.persistence.entity.TeacherEntity;
 import com.capgemini.gradebook.persistence.repo.TeacherRepo;
 import com.capgemini.gradebook.service.TeacherService;
+import com.capgemini.gradebook.service.exception.BadRequestException;
+import com.capgemini.gradebook.service.exception.NotFoundException;
 
 @Service
 @Transactional
@@ -43,13 +46,15 @@ public class TeacherServiceImpl implements TeacherService {
   public TeacherEto findTeacherById(Long id) {
 
     final Optional<TeacherEntity> result = this.teacherRepository.findById(id);
-    //TODO IMPLEMENT: Throw new custom exception
-    return result.map(r -> TeacherMapper.mapToETO(r)).orElseThrow( ()-> new RuntimeException("Teacher not found " +
-        "exception"));
+    return result.map(r -> TeacherMapper.mapToETO(r)).orElseThrow( ()-> new NotFoundException("Teacher not found"));
   }
 
   @Override
   public TeacherEto save(TeacherEto newTeacher) {
+
+    if (newTeacher.getId() != null) {
+      throw new BadRequestException("Teacher with such id already exists");
+    }
 
     TeacherEntity teacherEntity = TeacherMapper.mapToEntity(newTeacher);
     teacherEntity = this.teacherRepository.save(teacherEntity);
@@ -57,9 +62,30 @@ public class TeacherServiceImpl implements TeacherService {
   }
 
   @Override
+  public TeacherEto updateTeacher(TeacherEto teacherEto, Long teacherId) {
+    String teacherFirstName = teacherEto.getFirstName();
+    String teacherLastName = teacherEto.getLastName();
+    TeacherEntity teacher =
+        this.teacherRepository.findById(teacherId).orElseThrow(() -> new IllegalStateException(
+        "Teacher with" +
+        " id " + teacherId + " doesn't exist"));
+
+    if(teacherFirstName != null && teacherFirstName.length() > 0 && !Objects.equals(teacher.getFirstName(), teacherFirstName)) {
+      teacher.setFirstName(teacherFirstName);
+    }
+
+    if(teacherLastName != null && teacherLastName.length() > 0 && !Objects.equals(teacher.getLastName(), teacherLastName)) {
+      teacher.setLastName(teacherLastName);
+    }
+    return TeacherMapper.mapToETO(teacher);
+  }
+
+  @Override
   public void delete(Long id) {
+    if(!teacherRepository.existsById(id)) {
+      throw new NotFoundException("Teacher with id " + id + " does not exists");
+    }
 
     this.teacherRepository.deleteById(id);
-
   }
 }
